pyproject.toml
[tool.poetry]
name = "salary-per-hour"
version = "0.1.0"
description = ""
authors = ["Nafisah Nurul Hakim <nafisahnrl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.2.1"
psycopg2-binary = "^2.9.9"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
